Day-10/requirements.txt
crewai>=0.28.0
langchain-openai>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0

